Split scenarios
