+++ aries_vcx/tests/test_credential_retrieval.rs
@@ -233,7 +233,7 @@ async fn test_agency_pool_case_for_proof_req_doesnt_matter_for_retrieve_creds() 
 async fn test_agency_pool_it_should_fail_to_select_credentials_for_predicate() {
     use aries_vcx::utils::devsetup::SetupPoolDirectory;
     use utils::{
-        scenarios::create_proof_request_data,
+        scenarios::prover_select_credentials,
         test_agent::{create_test_agent, create_test_agent_trustee},
     };
